Add selectize for authors
Authors are special because a book may have multiple authors, and
authors may author multiple books. The form element that you want
for this is a select dropdown that can be populated from a list of
known values (which you can then autocomplete) + the ability to add
new values. Selectize gives you this.

This could be potentially be a bit cleaner if I used the symfony
form entity collection stuff, but this works for now and I'd like to
move on.